d/control: record lz4 build dependency
Since Ubuntu Eoan the kernel compression was changed from GZIP to
LZ4, due to slightly faster load times vs. a 25% size increase
trade-off (e.g. 5.0 had ~ 8, this one has ~ 12 MB; *but* the initrd
stays roughly the same size, and that one is 5 times bigger anyway)

If we want to keep that is in the stars, but for now correctly
document the build-dependency to LZ4.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>