Do not build custom build profiles anymore
This change removes the custom build profiles from the CI pipeline.
Previously, this repository offered kernels that were specifically
compiled for a certain microarchitecture.

However, since the effect of such optimizations for the kernel is rather
neglible and due to the increased cost of maintaining several build profiles,
I have decided to stop offering these as pre-builts.