Disable UBSAN
The upstream Ubuntu kernel enables UBSAN on 5.14+ and 5.13.14+ builds, leading
to boot issues on servers depending on the hardware or module configuration.

With hundreds of different kernel modules and thousands different
configuration, this change in its current stage is high risk and questionable
in general.

These issues are impossible to predict and hard to troubleshoot.
Therefore, I believe it is worth disabling for now.

Resolves #164