Address ZFS race condition
This change adds a patch that reverts a change to the Linux kernel that
was intended to clean-up some internal behavior. Unfortunately, ZFS
currently relies on this behavior to perform correctly.

Without this patch, zvol links might not be created correctly due to a
race condition.

See #183 for more information about this issue.